Change local ID to long. Added function to get entity with specified ID.

+++ source/src/org/societies/android/platform/entity/Community.java
@@ -31,7 +31,7 @@ import android.net.Uri;
  */
 public class Community extends Entity {
 
-	private int id = ENTITY_DEFAULT_ID;
+	private long id = ENTITY_DEFAULT_ID;
 	
 	private String globalId;
 	private String name;
@@ -62,7 +62,7 @@ public class Community extends Entity {
 	
 	@Override
 	protected void populate(Cursor cursor) {
-		setId(				Entity.getInt(cursor, _ID));
+		setId(				Entity.getLong(cursor, _ID));
 		setGlobalId(		Entity.getString(cursor, GLOBAL_ID));
 		setName(			Entity.getString(cursor, NAME));
 		setOwnerId(			Entity.getString(cursor, OWNER_ID));
@@ -98,12 +98,12 @@ public class Community extends Entity {
 	}
 
 	@Override
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	
 	@Override
-	protected void setId(int id) {
+	protected void setId(long id) {
 		this.id = id;
 	}
