+++ source/src/org/societies/android/platform/entity/CommunityActivity.java
@@ -32,7 +32,7 @@ import static org.societies.android.api.cis.SocialContract.CommunityActivity.*;
  */
 public class CommunityActivity extends Entity {
 
-	private int id = ENTITY_DEFAULT_ID;
+	private long id = ENTITY_DEFAULT_ID;
 	
 	private String globalId;
 	private long feedOwnerId;
@@ -71,7 +71,7 @@ public class CommunityActivity extends Entity {
 	
 	@Override
 	protected void populate(Cursor cursor) {
-		setId(				Entity.getInt(cursor, _ID));
+		setId(				Entity.getLong(cursor, _ID));
 		setGlobalId(		Entity.getString(cursor, GLOBAL_ID));
 		setFeedOwnerId(		Entity.getLong(cursor, _ID_FEED_OWNER));
 		setActor(			Entity.getString(cursor, ACTOR));
@@ -129,12 +129,12 @@ public class CommunityActivity extends Entity {
 	}
 
 	@Override
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	
 	@Override
-	protected void setId(int id) {
+	protected void setId(long id) {
 		this.id = id;
 	}
