+++ source/src/org/societies/android/platform/entity/Person.java
@@ -31,7 +31,7 @@ import android.net.Uri;
  */
 public class Person extends Entity {
 	
-	private int id = ENTITY_DEFAULT_ID;
+	private long id = ENTITY_DEFAULT_ID;
 	
 	private String globalId;
 	private String name;
@@ -60,7 +60,7 @@ public class Person extends Entity {
 	
 	@Override
 	protected void populate(Cursor cursor) {
-		setId(				Entity.getInt(cursor, _ID));
+		setId(				Entity.getLong(cursor, _ID));
 		setGlobalId(		Entity.getString(cursor, GLOBAL_ID));
 		setName(			Entity.getString(cursor, NAME));
 		setDescription(		Entity.getString(cursor, DESCRIPTION));
@@ -94,12 +94,12 @@ public class Person extends Entity {
 	}
 	
 	@Override
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	
 	@Override
-	protected void setId(int id) {
+	protected void setId(long id) {
 		this.id = id;
 	}
