+++ source/src/org/societies/android/platform/entity/Relationship.java
@@ -31,7 +31,7 @@ import static org.societies.android.api.cis.SocialContract.Relationship.*;
  */
 public class Relationship extends Entity {
 
-	private int id = ENTITY_DEFAULT_ID;
+	private long id = ENTITY_DEFAULT_ID;
 	
 	private String globalId;
 	private String globalIdP1;
@@ -61,7 +61,7 @@ public class Relationship extends Entity {
 
 	@Override
 	protected void populate(Cursor cursor) {
-		setId(				Entity.getInt(cursor, _ID));
+		setId(				Entity.getLong(cursor, _ID));
 		setGlobalId(		Entity.getString(cursor, GLOBAL_ID));
 		setGlobalIdP1(		Entity.getString(cursor, GLOBAL_ID_P1));
 		setGlobalIdP2(		Entity.getString(cursor, GLOBAL_ID_P2));
@@ -95,12 +95,12 @@ public class Relationship extends Entity {
 	}
 	
 	@Override
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	
 	@Override
-	protected void setId(int id) {
+	protected void setId(long id) {
 		this.id = id;
 	}
