+++ source/src/org/societies/android/platform/entity/Sharing.java
@@ -31,7 +31,7 @@ import static org.societies.android.api.cis.SocialContract.Sharing.*;
  */
 public class Sharing extends Entity {
 
-	private int id = ENTITY_DEFAULT_ID;
+	private long id = ENTITY_DEFAULT_ID;
 	
 	private String globalId;
 	private String globalIdService;
@@ -62,7 +62,7 @@ public class Sharing extends Entity {
 	
 	@Override
 	protected void populate(Cursor cursor) {
-		setId(					Entity.getInt(cursor, _ID));
+		setId(					Entity.getLong(cursor, _ID));
 		setGlobalId(			Entity.getString(cursor, GLOBAL_ID));
 		setGlobalIdService(		Entity.getString(cursor, GLOBAL_ID_SERVICE));
 		setGlobalIdOwner(		Entity.getString(cursor, GLOBAL_ID_OWNER));
@@ -98,12 +98,12 @@ public class Sharing extends Entity {
 	}
 	
 	@Override
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	
 	@Override
-	protected void setId(int id) {
+	protected void setId(long id) {
 		this.id = id;
 	}
